+++ docs/main.js
@@ -1,1 +1,1 @@
-(()=>{var e={616:e=>{"use strict";e.exports=function(e){var n=[];return n.toString=function(){return this.map((function(n){var t="",r=void 0!==n[5];return n[4]&&(t+="@supports (".concat(n[4],") {")),n[2]&&(t+="@media ".concat(n[2]," {")),r&&(t+="@layer".concat(n[5].length>0?" ".concat(n[5]):""," {")),t+=e(n),r&&(t+="}"),n[2]&&(t+="}"),n[4]&&(t+="}"),t})).join("")},n.i=function(e,t,r,o,i){"string"==typeof e&&(e=[[null,e,void 0]]);var a={};if(r)for(var s=0;s<this.length;s++){var c=this[s][0];null!=c&&(a[c]=!0)}for(var l=0;l<e.length;l++){var d=[].concat(e[l]);r&&a[d[0]]||(void 0!==i&&(void 0===d[5]||(d[1]="@layer".concat(d[5].length>0?" ".concat(d[5]):""," {").concat(d[1],"}")),d[5]=i),t&&(d[2]?(d[1]="@media ".concat(d[2]," {").concat(d[1],"}"),d[2]=t):d[2]=t),o&&(d[4]?(d[1]="@supports (".concat(d[4],") {").concat(d[1],"}"),d[4]=o):d[4]="".concat(o)),n.push(d))}},n}},676:e=>{"use strict";e.exports=function(e,n){return n||(n={}),e?(e=String(e.__esModule?e.default:e),/^['"].*['"]$/.test(e)&&(e=e.slice(1,-1)),n.hash&&(e+=n.hash),/["'() \t\n]|(%20)/.test(e)||n.needQuotes?'"'.concat(e.replace(/"/g,'\\"').replace(/\n/g,"\\n"),'"'):e):e}},515:e=>{"use strict";e.exports=function(e){return e[1]}},264:(e,n,t)=>{t(271),e.exports=self.fetch.bind(self)},271:(e,n,t)=>{"use strict";function r(e){return r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e},r(e)}t.r(n),t.d(n,{DOMException:()=>A,Headers:()=>h,Request:()=>E,Response:()=>B,fetch:()=>I});var o="undefined"!=typeof globalThis&&globalThis||"undefined"!=typeof self&&self||void 0!==o&&o,i="URLSearchParams"in o,a="Symbol"in o&&"iterator"in Symbol,s="FileReader"in o&&"Blob"in o&&function(){try{return new Blob,!0}catch(e){return!1}}(),c="FormData"in o,l="ArrayBuffer"in o;if(l)var d=["[object Int8Array]","[object Uint8Array]","[object Uint8ClampedArray]","[object Int16Array]","[object Uint16Array]","[object Int32Array]","[object Uint32Array]","[object Float32Array]","[object Float64Array]"],u=ArrayBuffer.isView||function(e){return e&&d.indexOf(Object.prototype.toString.call(e))>-1};function p(e){if("string"!=typeof e&&(e=String(e)),/[^a-z0-9\-#$%&'*+.^_`|~!]/i.test(e)||""===e)throw new TypeError('Invalid character in header field name: "'+e+'"');return e.toLowerCase()}function f(e){return"string"!=typeof e&&(e=String(e)),e}function y(e){var n={next:function(){var n=e.shift();return{done:void 0===n,value:n}}};return a&&(n[Symbol.iterator]=function(){return n}),n}function h(e){this.map={},e instanceof h?e.forEach((function(e,n){this.append(n,e)}),this):Array.isArray(e)?e.forEach((function(e){this.append(e[0],e[1])}),this):e&&Object.getOwnPropertyNames(e).forEach((function(n){this.append(n,e[n])}),this)}function m(e){if(e.bodyUsed)return Promise.reject(new TypeError("Already read"));e.bodyUsed=!0}function b(e){return new Promise((function(n,t){e.onload=function(){n(e.result)},e.onerror=function(){t(e.error)}}))}function g(e){var n=new FileReader,t=b(n);return n.readAsArrayBuffer(e),t}function x(e){if(e.slice)return e.slice(0);var n=new Uint8Array(e.byteLength);return n.set(new Uint8Array(e)),n.buffer}function v(){return this.bodyUsed=!1,this._initBody=function(e){var n;this.bodyUsed=this.bodyUsed,this._bodyInit=e,e?"string"==typeof e?this._bodyText=e:s&&Blob.prototype.isPrototypeOf(e)?this._bodyBlob=e:c&&FormData.prototype.isPrototypeOf(e)?this._bodyFormData=e:i&&URLSearchParams.prototype.isPrototypeOf(e)?this._bodyText=e.toString():l&&s&&(n=e)&&DataView.prototype.isPrototypeOf(n)?(this._bodyArrayBuffer=x(e.buffer),this._bodyInit=new Blob([this._bodyArrayBuffer])):l&&(ArrayBuffer.prototype.isPrototypeOf(e)||u(e))?this._bodyArrayBuffer=x(e):this._bodyText=e=Object.prototype.toString.call(e):this._bodyText="",this.headers.get("content-type")||("string"==typeof e?this.headers.set("content-type","text/plain;charset=UTF-8"):this._bodyBlob&&this._bodyBlob.type?this.headers.set("content-type",this._bodyBlob.type):i&&URLSearchParams.prototype.isPrototypeOf(e)&&this.headers.set("content-type","application/x-www-form-urlencoded;charset=UTF-8"))},s&&(this.blob=function(){var e=m(this);if(e)return e;if(this._bodyBlob)return Promise.resolve(this._bodyBlob);if(this._bodyArrayBuffer)return Promise.resolve(new Blob([this._bodyArrayBuffer]));if(this._bodyFormData)throw new Error("could not read FormData body as blob");return Promise.resolve(new Blob([this._bodyText]))},this.arrayBuffer=function(){return this._bodyArrayBuffer?m(this)||(ArrayBuffer.isView(this._bodyArrayBuffer)?Promise.resolve(this._bodyArrayBuffer.buffer.slice(this._bodyArrayBuffer.byteOffset,this._bodyArrayBuffer.byteOffset+this._bodyArrayBuffer.byteLength)):Promise.resolve(this._bodyArrayBuffer)):this.blob().then(g)}),this.text=function(){var e,n,t,r=m(this);if(r)return r;if(this._bodyBlob)return e=this._bodyBlob,t=b(n=new FileReader),n.readAsText(e),t;if(this._bodyArrayBuffer)return Promise.resolve(function(e){for(var n=new Uint8Array(e),t=new Array(n.length),r=0;r<n.length;r++)t[r]=String.fromCharCode(n[r]);return t.join("")}(this._bodyArrayBuffer));if(this._bodyFormData)throw new Error("could not read FormData body as text");return Promise.resolve(this._bodyText)},c&&(this.formData=function(){return this.text().then(S)}),this.json=function(){return this.text().then(JSON.parse)},this}h.prototype.append=function(e,n){e=p(e),n=f(n);var t=this.map[e];this.map[e]=t?t+", "+n:n},h.prototype.delete=function(e){delete this.map[p(e)]},h.prototype.get=function(e){return e=p(e),this.has(e)?this.map[e]:null},h.prototype.has=function(e){return this.map.hasOwnProperty(p(e))},h.prototype.set=function(e,n){this.map[p(e)]=f(n)},h.prototype.forEach=function(e,n){for(var t in this.map)this.map.hasOwnProperty(t)&&e.call(n,this.map[t],t,this)},h.prototype.keys=function(){var e=[];return this.forEach((function(n,t){e.push(t)})),y(e)},h.prototype.values=function(){var e=[];return this.forEach((function(n){e.push(n)})),y(e)},h.prototype.entries=function(){var e=[];return this.forEach((function(n,t){e.push([t,n])})),y(e)},a&&(h.prototype[Symbol.iterator]=h.prototype.entries);var w=["DELETE","GET","HEAD","OPTIONS","POST","PUT"];function E(e,n){if(!(this instanceof E))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');var t,r,o=(n=n||{}).body;if(e instanceof E){if(e.bodyUsed)throw new TypeError("Already read");this.url=e.url,this.credentials=e.credentials,n.headers||(this.headers=new h(e.headers)),this.method=e.method,this.mode=e.mode,this.signal=e.signal,o||null==e._bodyInit||(o=e._bodyInit,e.bodyUsed=!0)}else this.url=String(e);if(this.credentials=n.credentials||this.credentials||"same-origin",!n.headers&&this.headers||(this.headers=new h(n.headers)),this.method=(r=(t=n.method||this.method||"GET").toUpperCase(),w.indexOf(r)>-1?r:t),this.mode=n.mode||this.mode||null,this.signal=n.signal||this.signal,this.referrer=null,("GET"===this.method||"HEAD"===this.method)&&o)throw new TypeError("Body not allowed for GET or HEAD requests");if(this._initBody(o),!("GET"!==this.method&&"HEAD"!==this.method||"no-store"!==n.cache&&"no-cache"!==n.cache)){var i=/([?&])_=[^&]*/;i.test(this.url)?this.url=this.url.replace(i,"$1_="+(new Date).getTime()):this.url+=(/\?/.test(this.url)?"&":"?")+"_="+(new Date).getTime()}}function S(e){var n=new FormData;return e.trim().split("&").forEach((function(e){if(e){var t=e.split("="),r=t.shift().replace(/\+/g," "),o=t.join("=").replace(/\+/g," ");n.append(decodeURIComponent(r),decodeURIComponent(o))}})),n}function B(e,n){if(!(this instanceof B))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');n||(n={}),this.type="default",this.status=void 0===n.status?200:n.status,this.ok=this.status>=200&&this.status<300,this.statusText=void 0===n.statusText?"":""+n.statusText,this.headers=new h(n.headers),this.url=n.url||"",this._initBody(e)}E.prototype.clone=function(){return new E(this,{body:this._bodyInit})},v.call(E.prototype),v.call(B.prototype),B.prototype.clone=function(){return new B(this._bodyInit,{status:this.status,statusText:this.statusText,headers:new h(this.headers),url:this.url})},B.error=function(){var e=new B(null,{status:0,statusText:""});return e.type="error",e};var T=[301,302,303,307,308];B.redirect=function(e,n){if(-1===T.indexOf(n))throw new RangeError("Invalid status code");return new B(null,{status:n,headers:{location:e}})};var A=o.DOMException;try{new A}catch(e){(A=function(e,n){this.message=e,this.name=n;var t=Error(e);this.stack=t.stack}).prototype=Object.create(Error.prototype),A.prototype.constructor=A}function I(e,n){return new Promise((function(t,i){var a=new E(e,n);if(a.signal&&a.signal.aborted)return i(new A("Aborted","AbortError"));var c=new XMLHttpRequest;function d(){c.abort()}c.onload=function(){var e,n,r={status:c.status,statusText:c.statusText,headers:(e=c.getAllResponseHeaders()||"",n=new h,e.replace(/\r?\n[\t ]+/g," ").split("\r").map((function(e){return 0===e.indexOf("\n")?e.substr(1,e.length):e})).forEach((function(e){var t=e.split(":"),r=t.shift().trim();if(r){var o=t.join(":").trim();n.append(r,o)}})),n)};r.url="responseURL"in c?c.responseURL:r.headers.get("X-Request-URL");var o="response"in c?c.response:c.responseText;setTimeout((function(){t(new B(o,r))}),0)},c.onerror=function(){setTimeout((function(){i(new TypeError("Network request failed"))}),0)},c.ontimeout=function(){setTimeout((function(){i(new TypeError("Network request failed"))}),0)},c.onabort=function(){setTimeout((function(){i(new A("Aborted","AbortError"))}),0)},c.open(a.method,function(e){try{return""===e&&o.location.href?o.location.href:e}catch(n){return e}}(a.url),!0),"include"===a.credentials?c.withCredentials=!0:"omit"===a.credentials&&(c.withCredentials=!1),"responseType"in c&&(s?c.responseType="blob":l&&a.headers.get("Content-Type")&&-1!==a.headers.get("Content-Type").indexOf("application/octet-stream")&&(c.responseType="arraybuffer")),!n||"object"!==r(n.headers)||n.headers instanceof h?a.headers.forEach((function(e,n){c.setRequestHeader(n,e)})):Object.getOwnPropertyNames(n.headers).forEach((function(e){c.setRequestHeader(e,f(n.headers[e]))})),a.signal&&(a.signal.addEventListener("abort",d),c.onreadystatechange=function(){4===c.readyState&&a.signal.removeEventListener("abort",d)}),c.send(void 0===a._bodyInit?null:a._bodyInit)}))}I.polyfill=!0,o.fetch||(o.fetch=I,o.Headers=h,o.Request=E,o.Response=B)},426:(e,n,t)=>{"use strict";t.d(n,{Z:()=>g});var r=t(515),o=t.n(r),i=t(616),a=t.n(i),s=t(676),c=t.n(s),l=new URL(t(67),t.b),d=new URL(t(668),t.b),u=new URL(t(468),t.b),p=new URL(t(900),t.b),f=a()(o()),y=c()(l),h=c()(d),m=c()(u),b=c()(p);f.push([e.id,"* {\n    margin: 0;\n    padding: 0;\n}\n\nbody {\n    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;\n    background: linear-gradient(to right, #ec2F4B, #009fff);\n    background-attachment: fixed;\n    background-position: center;\n    background-size: cover;\n    padding: 30px;\n}\n\n\n\n.titulo h1 {\n    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;\n    color: red;\n    font-size: 35px;\n    font-weight: normal;\n    font-style: italic;\n    text-shadow: 7px 4px 5px blue;\n    text-align: center;\n\n}\n\n.darkLayout {\n    color: white !important;\n    background: #2E2E2E;\n\n}\n\n.darkLayout:hover {\n    color: blueviolet\n}\n    \nh1:hover {\n    color: aliceblue;\n}\n\n\n\n.imagen-body img {\n    width: 300px;\n    align-items: center;\n    padding: 40px;\n    margin: auto;\n    display: block;\n    border-radius: 100px;\n}\n\n.busqueda name {\n    text-align: center;\n\n}\n\n.busqueda {\n    width: 100%;\n    max-width: 600px;\n    margin: 0 auto;\n    display: flex;\n    flex-direction: column;\n    justify-content: center;\n    align-items: center;\n    font-size: larger;\n    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;\n    color: aliceblue;\n}\n\n#enviar:hover,\nbutton {\n\n    color: rgb(9, 8, 8);\n    background-color: #d3681b;\n    border-radius: 5px;\n    max-width: 600px;\n    text-align: center;\n    padding: 10px;\n    font-family: helvetica;\n    font-weight: 300;\n    font-size: 25px;\n    border-radius: 15px;\n    border: 3px double #651800;\n    text-decoration: none;\n    margin: 40px;\n\n}\n#enviar:hover {\n    color: white;\n\n}\n\n#user-input {\n    border: 4px solid rgb(0, 255, 204);\n    width: 50%;\n    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;\n\n}\n\n.flex-container {\n    display: flex;\n    padding: 30px 0;\n    align-items: center;\n    text-align: center;\n    width: 100%;\n    height: 200px;\n    border: 2px solid red;\n    justify-content: space-between;\n\n}\n\n.caja {\n    display: flex;\n    align-items: center;\n    justify-content: center;\n    width: 250px;\n    height: 250px;\n    background-size: contain;\n    background-repeat: no-repeat;\n    font-size: 25px;\n    border-radius: 125px;\n    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;\n    font-weight: 600;\n    text-shadow: 7px 4px 5px rgb(0, 255, 8);\n    min-width: 250px;\n    max-width: 250px;\n    min-height: 250px;\n    max-height: 250px;\n\n\n}\n\n.caja1 {\n    background-image: url("+y+");\n    background-size: cover;\n    margin: 10px;\n    color: #f07110;\n    text-shadow: 7px 4px 5px rgb(0, 255, 4);\n\n\n}\n\n.caja2 {\n    background-image: url("+h+");\n    background-size: cover;\n    margin: 10px;\n    color: #080700;\n    align-items: center;\n\n}\n\n.caja3 {\n    background-image: url("+m+");\n    background-size: cover;\n    margin: 10px;\n    color: rgb(62, 221, 13);\n    text-shadow: 7px 4px 5px blue;\n}\n\n.caja4 {\n    background-image: url("+b+");\n    background-size: cover;\n    margin: 10px;\n    color: white;\n\n}\n\n.busqueda button {\n    cursor: pointer;\n\n}\n\n#botonMostrar button {\n    max-width: 600px;\n    text-align: center;\n    padding: 10px;\n    font-family: helvetica;\n    font-weight: 300;\n    font-size: 25px;\n    color: #080700;\n    background-color: #d3681b;\n    border-radius: 15px;\n    border: 3px double #651800;\n    text-decoration: none;\n    margin: 40px;\n\n}\n\n#botonMostrar button {\n    cursor: nw-resize;\n}\n\n#botonMostrar :hover {\n    color: white;\n}\n\n.borrar-ingredientes {\n    padding: 30px;\n    margin: 20px;\n    text-align: left;\n    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;\n\n}\n\n#imagen-bebida {\n    width: 400px;\n    height: 400px;\n}\n\n@media only screen and (max-width: 1300px) {\n    .flex-container {\n        border: 2px solid rgb(0, 34, 255);\n    }\n\n    .caja {\n        border: 4px solid rgb(239, 246, 238);\n        max-width: 180px;\n        min-width: 180px;\n        max-height: 180px;\n        min-height: 180px;\n\n    }\n}\n\n@media only screen and (max-width: 900px) {\n    .flex-container {\n        border: 2px solid rgb(0, 34, 255);\n    }\n\n    .caja {\n        border: 4px solid rgb(236, 148, 7);\n        max-width: 110px;\n        min-width: 110px;\n        max-height: 110px;\n        min-height: 110px;\n\n    }\n}\n\n\n@media only screen and (max-width: 578px) {\n    .flex-container {\n        /* flex-direction: row; ------asi se pone en columna o fila */\n        border: 2px solid red;\n    }\n\n    .caja {\n        border: 4px solid rgb(221, 0, 255);\n        max-width: 60px;\n        min-width: 60px;\n        max-height: 60px;\n        min-height: 60px;\n        font-size: 15px;\n\n    }\n}\n\n.parrafo-instrucciones {\n    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;\n    text-align: left;\n    margin: 20px;\n}\n\n.class-old-image {\n    border-radius: 20px;\n    border: 5px solid red;\n    width: 100%;\n}\n\n#ingredientes-id {\n    font-size: 20px;\n    text-align: left;\n}\n\n.borrar-ingredientes {\n    margin: 2px;\n    font-size: 16px;\n    text-align: left;\n}\n\n#instructions-id {\n    font-size: 16px;\n    text-align: left;\n}\n\n.flex-container-ddp {\n    display: flex;\n    flex-direction: column;\n    padding: 10px;\n    border: 4px solid rgb(255, 34, 0);\n\n}\n\n/* aqui se selecciona especificamente el o los elementos\n.flex-container-ddp div {\n    display: flex;\n    padding: 10px;\n    margin: 5px;\n    flex-grow: 2px;\n} */\n\n.flex-container-info-bebida {\n    display: flex;\n    flex-direction: row;\n    border: 4px solid greenyellow;\n    justify-content: space-around;\n    line-height: 20px;\n    text-align: center;\n\n}\n\n.checkboxBorder{\n    text-align: right;\n\n}\n\n/* @media only screen and (max-width: 390px) {\n    .caja {\n        flex-direction: column;\n    }\n} */",""]);const g=f},379:e=>{"use strict";var n=[];function t(e){for(var t=-1,r=0;r<n.length;r++)if(n[r].identifier===e){t=r;break}return t}function r(e,r){for(var i={},a=[],s=0;s<e.length;s++){var c=e[s],l=r.base?c[0]+r.base:c[0],d=i[l]||0,u="".concat(l," ").concat(d);i[l]=d+1;var p=t(u),f={css:c[1],media:c[2],sourceMap:c[3],supports:c[4],layer:c[5]};if(-1!==p)n[p].references++,n[p].updater(f);else{var y=o(f,r);r.byIndex=s,n.splice(s,0,{identifier:u,updater:y,references:1})}a.push(u)}return a}function o(e,n){var t=n.domAPI(n);return t.update(e),function(n){if(n){if(n.css===e.css&&n.media===e.media&&n.sourceMap===e.sourceMap&&n.supports===e.supports&&n.layer===e.layer)return;t.update(e=n)}else t.remove()}}e.exports=function(e,o){var i=r(e=e||[],o=o||{});return function(e){e=e||[];for(var a=0;a<i.length;a++){var s=t(i[a]);n[s].references--}for(var c=r(e,o),l=0;l<i.length;l++){var d=t(i[l]);0===n[d].references&&(n[d].updater(),n.splice(d,1))}i=c}}},569:e=>{"use strict";var n={};e.exports=function(e,t){var r=function(e){if(void 0===n[e]){var t=document.querySelector(e);if(window.HTMLIFrameElement&&t instanceof window.HTMLIFrameElement)try{t=t.contentDocument.head}catch(e){t=null}n[e]=t}return n[e]}(e);if(!r)throw new Error("Couldn't find a style target. This probably means that the value for the 'insert' parameter is invalid.");r.appendChild(t)}},216:e=>{"use strict";e.exports=function(e){var n=document.createElement("style");return e.setAttributes(n,e.attributes),e.insert(n,e.options),n}},565:(e,n,t)=>{"use strict";e.exports=function(e){var n=t.nc;n&&e.setAttribute("nonce",n)}},795:e=>{"use strict";e.exports=function(e){var n=e.insertStyleElement(e);return{update:function(t){!function(e,n,t){var r="";t.supports&&(r+="@supports (".concat(t.supports,") {")),t.media&&(r+="@media ".concat(t.media," {"));var o=void 0!==t.layer;o&&(r+="@layer".concat(t.layer.length>0?" ".concat(t.layer):""," {")),r+=t.css,o&&(r+="}"),t.media&&(r+="}"),t.supports&&(r+="}");var i=t.sourceMap;i&&"undefined"!=typeof btoa&&(r+="\n/*# sourceMappingURL=data:application/json;base64,".concat(btoa(unescape(encodeURIComponent(JSON.stringify(i))))," */")),n.styleTagTransform(r,e,n.options)}(n,e,t)},remove:function(){!function(e){if(null===e.parentNode)return!1;e.parentNode.removeChild(e)}(n)}}}},589:e=>{"use strict";e.exports=function(e,n){if(n.styleSheet)n.styleSheet.cssText=e;else{for(;n.firstChild;)n.removeChild(n.firstChild);n.appendChild(document.createTextNode(e))}}},668:(e,n,t)=>{"use strict";e.exports=t.p+"8e343a5be68c1b10399b.jpeg"},468:(e,n,t)=>{"use strict";e.exports=t.p+"6a63eb35f09099a0346c.jpeg"},900:(e,n,t)=>{"use strict";e.exports=t.p+"b02d86611cea02418d43.jpeg"},67:(e,n,t)=>{"use strict";e.exports=t.p+"8131275090515d5863dc.jpeg"}},n={};function t(r){var o=n[r];if(void 0!==o)return o.exports;var i=n[r]={id:r,exports:{}};return e[r](i,i.exports,t),i.exports}t.m=e,t.n=e=>{var n=e&&e.__esModule?()=>e.default:()=>e;return t.d(n,{a:n}),n},t.d=(e,n)=>{for(var r in n)t.o(n,r)&&!t.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:n[r]})},t.g=function(){if("object"==typeof globalThis)return globalThis;try{return this||new Function("return this")()}catch(e){if("object"==typeof window)return window}}(),t.o=(e,n)=>Object.prototype.hasOwnProperty.call(e,n),t.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},(()=>{var e;t.g.importScripts&&(e=t.g.location+"");var n=t.g.document;if(!e&&n&&(n.currentScript&&(e=n.currentScript.src),!e)){var r=n.getElementsByTagName("script");r.length&&(e=r[r.length-1].src)}if(!e)throw new Error("Automatic publicPath is not supported in this browser");e=e.replace(/#.*$/,"").replace(/\?.*$/,"").replace(/\/[^\/]+$/,"/"),t.p=e})(),t.b=document.baseURI||self.location.href,t.nc=void 0,(()=>{"use strict";var e=t(379),n=t.n(e),r=t(795),o=t.n(r),i=t(569),a=t.n(i),s=t(565),c=t.n(s),l=t(216),d=t.n(l),u=t(589),p=t.n(u),f=t(426),y={};function h(e,n){(null==n||n>e.length)&&(n=e.length);for(var t=0,r=new Array(n);t<n;t++)r[t]=e[t];return r}y.styleTagTransform=p(),y.setAttributes=c(),y.insert=a().bind(null,"head"),y.domAPI=o(),y.insertStyleElement=d(),n()(f.Z,y),f.Z&&f.Z.locals&&f.Z.locals;var m=t(264),b=document.getElementById("busqueda"),g=document.getElementById("enviar"),x=document.getElementById("mostrar"),v=document.getElementById("instructions-id"),w=document.getElementById("ingredientes-id"),E=document.getElementById("imagenDiv"),S=document.getElementById("infoBebidaDiv");x.style.display="none",v.style.display="none",w.style.display="none",S.style.display="none",document.querySelector(".cbox1").addEventListener("change",(function(e){e.target.checked?(document.querySelector(".flex-container").style.border="none",document.getElementById("infoBebidaDiv").style.border="none",document.querySelector(".flex-container-ddp").style.border="none",document.getElementById("user-input").style.border="none",document.querySelector("h1").classList.add("darkLayout"),document.querySelector("body").classList.add("darkLayout")):(document.querySelector(".flex-container").style.border="2px solid rgb(0, 34, 255)",document.getElementById("infoBebidaDiv").style.border="4px solid greenyellow",document.querySelector(".flex-container-ddp").style.border="4px solid rgb(255, 34, 0)",document.getElementById("user-input").style.border="4px solid rgb(0, 255, 204)",document.querySelector("h1").classList.remove("darkLayout"),document.querySelector("body").classList.remove("darkLayout"))}));var B=document.getElementById("user-input");g.addEventListener("click",(function(){var e;e=B.value,m("https://www.thecocktaildb.com/api/json/v1/1/search.php?s="+e).then((function(e){return e.json()})).then((function(e){if(document.getElementsByClassName("drink-type"))for(var n=document.getElementsByClassName("drink-type"),t=n.length-1;t>=0;t--)n[t].remove();var r,o=function(e,n){var t="undefined"!=typeof Symbol&&e[Symbol.iterator]||e["@@iterator"];if(!t){if(Array.isArray(e)||(t=function(e,n){if(e){if("string"==typeof e)return h(e,n);var t=Object.prototype.toString.call(e).slice(8,-1);return"Object"===t&&e.constructor&&(t=e.constructor.name),"Map"===t||"Set"===t?Array.from(e):"Arguments"===t||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(t)?h(e,n):void 0}}(e))||n&&e&&"number"==typeof e.length){t&&(e=t);var r=0,o=function(){};return{s:o,n:function(){return r>=e.length?{done:!0}:{done:!1,value:e[r++]}},e:function(e){throw e},f:o}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var i,a=!0,s=!1;return{s:function(){t=t.call(e)},n:function(){var e=t.next();return a=e.done,e},e:function(e){s=!0,i=e},f:function(){try{a||null==t.return||t.return()}finally{if(s)throw i}}}}(e.drinks);try{var i=function(){var e=r.value,n=document.createElement("p");n.innerText=e.strDrink,n.classList.add("drink-type"),b.append(n);var t=document.querySelector(".bar");n.addEventListener("click",(function(){console.log(n.innerText),document.querySelector(".flex-container").style.display="none",document.querySelector(".busqueda").style.display="none",x.style.display="block",console.log(e,"---");var r=document.createElement("img");if(r.classList.add("class-old-image"),r.setAttribute("src",e.strDrinkThumb),document.getElementsByClassName("class-old-image"))for(var o=document.getElementsByClassName("class-old-image"),i=o.length-1;i>=0;i--)o[i].remove();if(document.getElementsByClassName("parrafo-instrucciones"))for(var a=document.getElementsByClassName("parrafo-instrucciones"),s=a.length-1;s>=0;s--)a[s].remove();var c=document.getElementsByClassName("borrar-ingredientes");if(c)for(var l=c.length-1;l>=0;l--)c[l].remove();var d=document.createElement("p");d.classList.add("parrafo-instrucciones"),d.innerText=e.strInstructions,v.append(d),r.setAttribute("id","imagen-bebida"),E.append(r),v.style.display="block",w.style.display="block",S.style.display="flex";var u=e.strIngredient1,p=e.strIngredient2,f=e.strIngredient3,y=e.strIngredient4,h=e.strIngredient5,m=e.strIngredient6,b=e.strIngredient7,g=(e.strIngredient8,e.strIngredient9,e.strIngredient10,e.strIngredient11,e.strIngredient12,e.strIngredient13,e.strIngredient14,e.strIngredient15,"".concat(null!==u?u+(null!==p?", ":"."):"").concat(null!==p?p+(null!==f?", ":"."):"").concat(null!==f?f+(null!==y?", ":"."):"").concat(null!==y?y+(null!==h?", ":"."):"").concat(null!==h?h+(null!==m?", ":"."):"").concat(null!==m?m+(null!==b?", ":"."):""));console.log(g);var B=document.createElement("p");B.classList.add("borrar-ingredientes"),B.innerText=g,w.append(B),t.innerText=e.strDrink})),x.addEventListener("click",(function(){document.querySelector(".flex-container").style.display="flex",document.querySelector(".busqueda").style.display="flex",x.style.display="none",document.querySelector(".class-old-image").style.display="none",document.querySelector(".parrafo-instrucciones").style.display="none",document.querySelector(".borrar-ingredientes").style.display="none",t.innerText="Pollos BAR",v.style.display="none",w.style.display="none",S.style.display="none";for(var e=document.getElementsByClassName("drink-type"),n=e.length-1;n>=0;n--)e[n].remove()}))};for(o.s();!(r=o.n()).done;)i()}catch(e){o.e(e)}finally{o.f()}console.log(e)}))}))})()})();
+(()=>{var e={616:e=>{"use strict";e.exports=function(e){var n=[];return n.toString=function(){return this.map((function(n){var t="",r=void 0!==n[5];return n[4]&&(t+="@supports (".concat(n[4],") {")),n[2]&&(t+="@media ".concat(n[2]," {")),r&&(t+="@layer".concat(n[5].length>0?" ".concat(n[5]):""," {")),t+=e(n),r&&(t+="}"),n[2]&&(t+="}"),n[4]&&(t+="}"),t})).join("")},n.i=function(e,t,r,o,i){"string"==typeof e&&(e=[[null,e,void 0]]);var a={};if(r)for(var s=0;s<this.length;s++){var c=this[s][0];null!=c&&(a[c]=!0)}for(var l=0;l<e.length;l++){var d=[].concat(e[l]);r&&a[d[0]]||(void 0!==i&&(void 0===d[5]||(d[1]="@layer".concat(d[5].length>0?" ".concat(d[5]):""," {").concat(d[1],"}")),d[5]=i),t&&(d[2]?(d[1]="@media ".concat(d[2]," {").concat(d[1],"}"),d[2]=t):d[2]=t),o&&(d[4]?(d[1]="@supports (".concat(d[4],") {").concat(d[1],"}"),d[4]=o):d[4]="".concat(o)),n.push(d))}},n}},676:e=>{"use strict";e.exports=function(e,n){return n||(n={}),e?(e=String(e.__esModule?e.default:e),/^['"].*['"]$/.test(e)&&(e=e.slice(1,-1)),n.hash&&(e+=n.hash),/["'() \t\n]|(%20)/.test(e)||n.needQuotes?'"'.concat(e.replace(/"/g,'\\"').replace(/\n/g,"\\n"),'"'):e):e}},515:e=>{"use strict";e.exports=function(e){return e[1]}},264:(e,n,t)=>{t(271),e.exports=self.fetch.bind(self)},271:(e,n,t)=>{"use strict";function r(e){return r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e},r(e)}t.r(n),t.d(n,{DOMException:()=>A,Headers:()=>h,Request:()=>E,Response:()=>B,fetch:()=>I});var o="undefined"!=typeof globalThis&&globalThis||"undefined"!=typeof self&&self||void 0!==o&&o,i="URLSearchParams"in o,a="Symbol"in o&&"iterator"in Symbol,s="FileReader"in o&&"Blob"in o&&function(){try{return new Blob,!0}catch(e){return!1}}(),c="FormData"in o,l="ArrayBuffer"in o;if(l)var d=["[object Int8Array]","[object Uint8Array]","[object Uint8ClampedArray]","[object Int16Array]","[object Uint16Array]","[object Int32Array]","[object Uint32Array]","[object Float32Array]","[object Float64Array]"],u=ArrayBuffer.isView||function(e){return e&&d.indexOf(Object.prototype.toString.call(e))>-1};function p(e){if("string"!=typeof e&&(e=String(e)),/[^a-z0-9\-#$%&'*+.^_`|~!]/i.test(e)||""===e)throw new TypeError('Invalid character in header field name: "'+e+'"');return e.toLowerCase()}function f(e){return"string"!=typeof e&&(e=String(e)),e}function y(e){var n={next:function(){var n=e.shift();return{done:void 0===n,value:n}}};return a&&(n[Symbol.iterator]=function(){return n}),n}function h(e){this.map={},e instanceof h?e.forEach((function(e,n){this.append(n,e)}),this):Array.isArray(e)?e.forEach((function(e){this.append(e[0],e[1])}),this):e&&Object.getOwnPropertyNames(e).forEach((function(n){this.append(n,e[n])}),this)}function m(e){if(e.bodyUsed)return Promise.reject(new TypeError("Already read"));e.bodyUsed=!0}function b(e){return new Promise((function(n,t){e.onload=function(){n(e.result)},e.onerror=function(){t(e.error)}}))}function g(e){var n=new FileReader,t=b(n);return n.readAsArrayBuffer(e),t}function x(e){if(e.slice)return e.slice(0);var n=new Uint8Array(e.byteLength);return n.set(new Uint8Array(e)),n.buffer}function v(){return this.bodyUsed=!1,this._initBody=function(e){var n;this.bodyUsed=this.bodyUsed,this._bodyInit=e,e?"string"==typeof e?this._bodyText=e:s&&Blob.prototype.isPrototypeOf(e)?this._bodyBlob=e:c&&FormData.prototype.isPrototypeOf(e)?this._bodyFormData=e:i&&URLSearchParams.prototype.isPrototypeOf(e)?this._bodyText=e.toString():l&&s&&(n=e)&&DataView.prototype.isPrototypeOf(n)?(this._bodyArrayBuffer=x(e.buffer),this._bodyInit=new Blob([this._bodyArrayBuffer])):l&&(ArrayBuffer.prototype.isPrototypeOf(e)||u(e))?this._bodyArrayBuffer=x(e):this._bodyText=e=Object.prototype.toString.call(e):this._bodyText="",this.headers.get("content-type")||("string"==typeof e?this.headers.set("content-type","text/plain;charset=UTF-8"):this._bodyBlob&&this._bodyBlob.type?this.headers.set("content-type",this._bodyBlob.type):i&&URLSearchParams.prototype.isPrototypeOf(e)&&this.headers.set("content-type","application/x-www-form-urlencoded;charset=UTF-8"))},s&&(this.blob=function(){var e=m(this);if(e)return e;if(this._bodyBlob)return Promise.resolve(this._bodyBlob);if(this._bodyArrayBuffer)return Promise.resolve(new Blob([this._bodyArrayBuffer]));if(this._bodyFormData)throw new Error("could not read FormData body as blob");return Promise.resolve(new Blob([this._bodyText]))},this.arrayBuffer=function(){return this._bodyArrayBuffer?m(this)||(ArrayBuffer.isView(this._bodyArrayBuffer)?Promise.resolve(this._bodyArrayBuffer.buffer.slice(this._bodyArrayBuffer.byteOffset,this._bodyArrayBuffer.byteOffset+this._bodyArrayBuffer.byteLength)):Promise.resolve(this._bodyArrayBuffer)):this.blob().then(g)}),this.text=function(){var e,n,t,r=m(this);if(r)return r;if(this._bodyBlob)return e=this._bodyBlob,t=b(n=new FileReader),n.readAsText(e),t;if(this._bodyArrayBuffer)return Promise.resolve(function(e){for(var n=new Uint8Array(e),t=new Array(n.length),r=0;r<n.length;r++)t[r]=String.fromCharCode(n[r]);return t.join("")}(this._bodyArrayBuffer));if(this._bodyFormData)throw new Error("could not read FormData body as text");return Promise.resolve(this._bodyText)},c&&(this.formData=function(){return this.text().then(S)}),this.json=function(){return this.text().then(JSON.parse)},this}h.prototype.append=function(e,n){e=p(e),n=f(n);var t=this.map[e];this.map[e]=t?t+", "+n:n},h.prototype.delete=function(e){delete this.map[p(e)]},h.prototype.get=function(e){return e=p(e),this.has(e)?this.map[e]:null},h.prototype.has=function(e){return this.map.hasOwnProperty(p(e))},h.prototype.set=function(e,n){this.map[p(e)]=f(n)},h.prototype.forEach=function(e,n){for(var t in this.map)this.map.hasOwnProperty(t)&&e.call(n,this.map[t],t,this)},h.prototype.keys=function(){var e=[];return this.forEach((function(n,t){e.push(t)})),y(e)},h.prototype.values=function(){var e=[];return this.forEach((function(n){e.push(n)})),y(e)},h.prototype.entries=function(){var e=[];return this.forEach((function(n,t){e.push([t,n])})),y(e)},a&&(h.prototype[Symbol.iterator]=h.prototype.entries);var w=["DELETE","GET","HEAD","OPTIONS","POST","PUT"];function E(e,n){if(!(this instanceof E))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');var t,r,o=(n=n||{}).body;if(e instanceof E){if(e.bodyUsed)throw new TypeError("Already read");this.url=e.url,this.credentials=e.credentials,n.headers||(this.headers=new h(e.headers)),this.method=e.method,this.mode=e.mode,this.signal=e.signal,o||null==e._bodyInit||(o=e._bodyInit,e.bodyUsed=!0)}else this.url=String(e);if(this.credentials=n.credentials||this.credentials||"same-origin",!n.headers&&this.headers||(this.headers=new h(n.headers)),this.method=(r=(t=n.method||this.method||"GET").toUpperCase(),w.indexOf(r)>-1?r:t),this.mode=n.mode||this.mode||null,this.signal=n.signal||this.signal,this.referrer=null,("GET"===this.method||"HEAD"===this.method)&&o)throw new TypeError("Body not allowed for GET or HEAD requests");if(this._initBody(o),!("GET"!==this.method&&"HEAD"!==this.method||"no-store"!==n.cache&&"no-cache"!==n.cache)){var i=/([?&])_=[^&]*/;i.test(this.url)?this.url=this.url.replace(i,"$1_="+(new Date).getTime()):this.url+=(/\?/.test(this.url)?"&":"?")+"_="+(new Date).getTime()}}function S(e){var n=new FormData;return e.trim().split("&").forEach((function(e){if(e){var t=e.split("="),r=t.shift().replace(/\+/g," "),o=t.join("=").replace(/\+/g," ");n.append(decodeURIComponent(r),decodeURIComponent(o))}})),n}function B(e,n){if(!(this instanceof B))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');n||(n={}),this.type="default",this.status=void 0===n.status?200:n.status,this.ok=this.status>=200&&this.status<300,this.statusText=void 0===n.statusText?"":""+n.statusText,this.headers=new h(n.headers),this.url=n.url||"",this._initBody(e)}E.prototype.clone=function(){return new E(this,{body:this._bodyInit})},v.call(E.prototype),v.call(B.prototype),B.prototype.clone=function(){return new B(this._bodyInit,{status:this.status,statusText:this.statusText,headers:new h(this.headers),url:this.url})},B.error=function(){var e=new B(null,{status:0,statusText:""});return e.type="error",e};var T=[301,302,303,307,308];B.redirect=function(e,n){if(-1===T.indexOf(n))throw new RangeError("Invalid status code");return new B(null,{status:n,headers:{location:e}})};var A=o.DOMException;try{new A}catch(e){(A=function(e,n){this.message=e,this.name=n;var t=Error(e);this.stack=t.stack}).prototype=Object.create(Error.prototype),A.prototype.constructor=A}function I(e,n){return new Promise((function(t,i){var a=new E(e,n);if(a.signal&&a.signal.aborted)return i(new A("Aborted","AbortError"));var c=new XMLHttpRequest;function d(){c.abort()}c.onload=function(){var e,n,r={status:c.status,statusText:c.statusText,headers:(e=c.getAllResponseHeaders()||"",n=new h,e.replace(/\r?\n[\t ]+/g," ").split("\r").map((function(e){return 0===e.indexOf("\n")?e.substr(1,e.length):e})).forEach((function(e){var t=e.split(":"),r=t.shift().trim();if(r){var o=t.join(":").trim();n.append(r,o)}})),n)};r.url="responseURL"in c?c.responseURL:r.headers.get("X-Request-URL");var o="response"in c?c.response:c.responseText;setTimeout((function(){t(new B(o,r))}),0)},c.onerror=function(){setTimeout((function(){i(new TypeError("Network request failed"))}),0)},c.ontimeout=function(){setTimeout((function(){i(new TypeError("Network request failed"))}),0)},c.onabort=function(){setTimeout((function(){i(new A("Aborted","AbortError"))}),0)},c.open(a.method,function(e){try{return""===e&&o.location.href?o.location.href:e}catch(n){return e}}(a.url),!0),"include"===a.credentials?c.withCredentials=!0:"omit"===a.credentials&&(c.withCredentials=!1),"responseType"in c&&(s?c.responseType="blob":l&&a.headers.get("Content-Type")&&-1!==a.headers.get("Content-Type").indexOf("application/octet-stream")&&(c.responseType="arraybuffer")),!n||"object"!==r(n.headers)||n.headers instanceof h?a.headers.forEach((function(e,n){c.setRequestHeader(n,e)})):Object.getOwnPropertyNames(n.headers).forEach((function(e){c.setRequestHeader(e,f(n.headers[e]))})),a.signal&&(a.signal.addEventListener("abort",d),c.onreadystatechange=function(){4===c.readyState&&a.signal.removeEventListener("abort",d)}),c.send(void 0===a._bodyInit?null:a._bodyInit)}))}I.polyfill=!0,o.fetch||(o.fetch=I,o.Headers=h,o.Request=E,o.Response=B)},426:(e,n,t)=>{"use strict";t.d(n,{Z:()=>g});var r=t(515),o=t.n(r),i=t(616),a=t.n(i),s=t(676),c=t.n(s),l=new URL(t(67),t.b),d=new URL(t(668),t.b),u=new URL(t(468),t.b),p=new URL(t(900),t.b),f=a()(o()),y=c()(l),h=c()(d),m=c()(u),b=c()(p);f.push([e.id,"* {\n    margin: 0;\n    padding: 0;\n}\n\nbody {\n    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;\n    background: linear-gradient(to right, #ec2F4B, #009fff);\n    background-attachment: fixed;\n    background-position: center;\n    background-size: cover;\n    padding: 30px;\n}\n\n\n\n.titulo h1 {\n    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;\n    color: red;\n    font-size: 35px;\n    font-weight: normal;\n    font-style: italic;\n    text-shadow: 7px 4px 5px blue;\n    text-align: center;\n\n}\n\n.darkLayout {\n    color: white !important;\n    background: #2E2E2E;\n\n}\n\n.darkLayout:hover {\n    color: blueviolet\n}\n    \nh1:hover {\n    color: aliceblue;\n}\n\n\n\n.imagen-body img {\n    width: 300px;\n    align-items: center;\n    padding: 40px;\n    margin: auto;\n    display: block;\n    border-radius: 100px;\n}\n\n.busqueda name {\n    text-align: center;\n\n}\n\n.busqueda {\n    width: 100%;\n    max-width: 600px;\n    margin: 0 auto;\n    display: flex;\n    flex-direction: column;\n    justify-content: center;\n    align-items: center;\n    font-size: larger;\n    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;\n    color: aliceblue;\n}\n\n#enviar:hover,\nbutton {\n\n    color: rgb(9, 8, 8);\n    background-color: #d3681b;\n    border-radius: 5px;\n    max-width: 600px;\n    text-align: center;\n    padding: 10px;\n    font-family: helvetica;\n    font-weight: 300;\n    font-size: 25px;\n    border-radius: 15px;\n    border: 3px double #651800;\n    text-decoration: none;\n    margin: 40px;\n\n}\n#enviar:hover {\n    color: white;\n\n}\n\n#user-input {\n    border: 4px solid rgb(0, 255, 204);\n    width: 50%;\n    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;\n\n}\n\n.flex-container {\n    display: flex;\n    padding: 30px 0;\n    align-items: center;\n    text-align: center;\n    width: 100%;\n    height: 200px;\n    border: 2px solid red;\n    justify-content: space-between;\n\n}\n\n.caja {\n    display: flex;\n    align-items: center;\n    justify-content: center;\n    width: 250px;\n    height: 250px;\n    background-size: contain;\n    background-repeat: no-repeat;\n    font-size: 25px;\n    border-radius: 125px;\n    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;\n    font-weight: 600;\n    text-shadow: 7px 4px 5px rgb(0, 255, 8);\n    min-width: 250px;\n    max-width: 250px;\n    min-height: 250px;\n    max-height: 250px;\n\n\n}\n\n.caja1 {\n    background-image: url("+y+");\n    background-size: cover;\n    margin: 10px;\n    color: #f07110;\n    text-shadow: 7px 4px 5px rgb(0, 255, 4);\n\n\n}\n\n.caja2 {\n    background-image: url("+h+");\n    background-size: cover;\n    margin: 10px;\n    color: #080700;\n    align-items: center;\n\n}\n\n.caja3 {\n    background-image: url("+m+");\n    background-size: cover;\n    margin: 10px;\n    color: rgb(62, 221, 13);\n    text-shadow: 7px 4px 5px blue;\n}\n\n.caja4 {\n    background-image: url("+b+");\n    background-size: cover;\n    margin: 10px;\n    color: white;\n\n}\n\n.busqueda button {\n    cursor: pointer;\n\n}\n\n#botonMostrar button {\n    max-width: 600px;\n    text-align: center;\n    padding: 10px;\n    font-family: helvetica;\n    font-weight: 300;\n    font-size: 25px;\n    color: #080700;\n    background-color: #d3681b;\n    border-radius: 15px;\n    border: 3px double #651800;\n    text-decoration: none;\n    margin: 40px;\n\n}\n\n#botonMostrar button {\n    cursor: nw-resize;\n}\n\n#botonMostrar :hover {\n    color: white;\n}\n\n.borrar-ingredientes {\n    padding: 30px;\n    margin: 20px;\n    text-align: left;\n    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;\n\n}\n\n#imagen-bebida {\n    width: 400px;\n    height: 400px;\n}\n\n@media only screen and (max-width: 1300px) {\n    .flex-container {\n        border: 2px solid rgb(0, 34, 255);\n    }\n\n    .caja {\n        border: 4px solid rgb(239, 246, 238);\n        max-width: 180px;\n        min-width: 180px;\n        max-height: 180px;\n        min-height: 180px;\n\n    }\n}\n\n@media only screen and (max-width: 900px) {\n    .flex-container {\n        border: 2px solid rgb(0, 34, 255);\n    }\n\n    .caja {\n        border: 4px solid rgb(236, 148, 7);\n        max-width: 110px;\n        min-width: 110px;\n        max-height: 110px;\n        min-height: 110px;\n\n    }\n}\n\n\n@media only screen and (max-width: 578px) {\n    .flex-container {\n        /* flex-direction: row; ------asi se pone en columna o fila */\n        border: 2px solid red;\n    }\n\n    .caja {\n        border: 4px solid rgb(221, 0, 255);\n        max-width: 60px;\n        min-width: 60px;\n        max-height: 60px;\n        min-height: 60px;\n        font-size: 15px;\n\n    }\n}\n\n.parrafo-instrucciones {\n    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;\n    text-align: left;\n    margin: 20px;\n}\n\n.class-old-image {\n    border-radius: 20px;\n    border: 5px solid red;\n    width: 100%;\n}\n\n#ingredientes-id {\n    font-size: 20px;\n    text-align: left;\n}\n\n.borrar-ingredientes {\n    margin: 2px;\n    font-size: 16px;\n    text-align: left;\n}\n\n#instructions-id {\n    font-size: 16px;\n    text-align: left;\n}\n\n.flex-container-ddp {\n    display: flex;\n    flex-direction: column;\n    padding: 10px;\n    border: 4px solid rgb(255, 34, 0);\n\n}\n\n/* aqui se selecciona especificamente el o los elementos\n.flex-container-ddp div {\n    display: flex;\n    padding: 10px;\n    margin: 5px;\n    flex-grow: 2px;\n} */\n\n.flex-container-info-bebida {\n    display: flex;\n    flex-direction: row;\n    border: 4px solid greenyellow;\n    justify-content: space-around;\n    line-height: 20px;\n    text-align: center;\n\n}\n\n.checkboxBorder{\n    text-align: right;\n\n}\n\n/* @media only screen and (max-width: 390px) {\n    .caja {\n        flex-direction: column;\n    }\n} */",""]);const g=f},379:e=>{"use strict";var n=[];function t(e){for(var t=-1,r=0;r<n.length;r++)if(n[r].identifier===e){t=r;break}return t}function r(e,r){for(var i={},a=[],s=0;s<e.length;s++){var c=e[s],l=r.base?c[0]+r.base:c[0],d=i[l]||0,u="".concat(l," ").concat(d);i[l]=d+1;var p=t(u),f={css:c[1],media:c[2],sourceMap:c[3],supports:c[4],layer:c[5]};if(-1!==p)n[p].references++,n[p].updater(f);else{var y=o(f,r);r.byIndex=s,n.splice(s,0,{identifier:u,updater:y,references:1})}a.push(u)}return a}function o(e,n){var t=n.domAPI(n);return t.update(e),function(n){if(n){if(n.css===e.css&&n.media===e.media&&n.sourceMap===e.sourceMap&&n.supports===e.supports&&n.layer===e.layer)return;t.update(e=n)}else t.remove()}}e.exports=function(e,o){var i=r(e=e||[],o=o||{});return function(e){e=e||[];for(var a=0;a<i.length;a++){var s=t(i[a]);n[s].references--}for(var c=r(e,o),l=0;l<i.length;l++){var d=t(i[l]);0===n[d].references&&(n[d].updater(),n.splice(d,1))}i=c}}},569:e=>{"use strict";var n={};e.exports=function(e,t){var r=function(e){if(void 0===n[e]){var t=document.querySelector(e);if(window.HTMLIFrameElement&&t instanceof window.HTMLIFrameElement)try{t=t.contentDocument.head}catch(e){t=null}n[e]=t}return n[e]}(e);if(!r)throw new Error("Couldn't find a style target. This probably means that the value for the 'insert' parameter is invalid.");r.appendChild(t)}},216:e=>{"use strict";e.exports=function(e){var n=document.createElement("style");return e.setAttributes(n,e.attributes),e.insert(n,e.options),n}},565:(e,n,t)=>{"use strict";e.exports=function(e){var n=t.nc;n&&e.setAttribute("nonce",n)}},795:e=>{"use strict";e.exports=function(e){var n=e.insertStyleElement(e);return{update:function(t){!function(e,n,t){var r="";t.supports&&(r+="@supports (".concat(t.supports,") {")),t.media&&(r+="@media ".concat(t.media," {"));var o=void 0!==t.layer;o&&(r+="@layer".concat(t.layer.length>0?" ".concat(t.layer):""," {")),r+=t.css,o&&(r+="}"),t.media&&(r+="}"),t.supports&&(r+="}");var i=t.sourceMap;i&&"undefined"!=typeof btoa&&(r+="\n/*# sourceMappingURL=data:application/json;base64,".concat(btoa(unescape(encodeURIComponent(JSON.stringify(i))))," */")),n.styleTagTransform(r,e,n.options)}(n,e,t)},remove:function(){!function(e){if(null===e.parentNode)return!1;e.parentNode.removeChild(e)}(n)}}}},589:e=>{"use strict";e.exports=function(e,n){if(n.styleSheet)n.styleSheet.cssText=e;else{for(;n.firstChild;)n.removeChild(n.firstChild);n.appendChild(document.createTextNode(e))}}},668:(e,n,t)=>{"use strict";e.exports=t.p+"8e343a5be68c1b10399b.jpeg"},468:(e,n,t)=>{"use strict";e.exports=t.p+"6a63eb35f09099a0346c.jpeg"},900:(e,n,t)=>{"use strict";e.exports=t.p+"b02d86611cea02418d43.jpeg"},67:(e,n,t)=>{"use strict";e.exports=t.p+"8131275090515d5863dc.jpeg"}},n={};function t(r){var o=n[r];if(void 0!==o)return o.exports;var i=n[r]={id:r,exports:{}};return e[r](i,i.exports,t),i.exports}t.m=e,t.n=e=>{var n=e&&e.__esModule?()=>e.default:()=>e;return t.d(n,{a:n}),n},t.d=(e,n)=>{for(var r in n)t.o(n,r)&&!t.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:n[r]})},t.g=function(){if("object"==typeof globalThis)return globalThis;try{return this||new Function("return this")()}catch(e){if("object"==typeof window)return window}}(),t.o=(e,n)=>Object.prototype.hasOwnProperty.call(e,n),t.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},(()=>{var e;t.g.importScripts&&(e=t.g.location+"");var n=t.g.document;if(!e&&n&&(n.currentScript&&(e=n.currentScript.src),!e)){var r=n.getElementsByTagName("script");r.length&&(e=r[r.length-1].src)}if(!e)throw new Error("Automatic publicPath is not supported in this browser");e=e.replace(/#.*$/,"").replace(/\?.*$/,"").replace(/\/[^\/]+$/,"/"),t.p=e})(),t.b=document.baseURI||self.location.href,t.nc=void 0,(()=>{"use strict";var e=t(379),n=t.n(e),r=t(795),o=t.n(r),i=t(569),a=t.n(i),s=t(565),c=t.n(s),l=t(216),d=t.n(l),u=t(589),p=t.n(u),f=t(426),y={};y.styleTagTransform=p(),y.setAttributes=c(),y.insert=a().bind(null,"head"),y.domAPI=o(),y.insertStyleElement=d(),n()(f.Z,y),f.Z&&f.Z.locals&&f.Z.locals;var h={listOfDrinks:document.getElementById("busqueda"),loadMoreBtn:document.getElementById("enviar"),regresar:document.getElementById("mostrar"),instructionsDiv:document.getElementById("instructions-id"),imagenDiv:document.getElementById("imagenDiv"),ingredientsDiv:document.getElementById("ingredientes-id"),flexContainer:document.getElementById("infoBebidaDiv"),titulo:document.querySelector(".bar")},m=h.regresar,b=h.instructionsDiv,g=h.imagenDiv,x=h.ingredientsDiv,v=h.flexContainer,w=h.titulo;function E(e,n){(null==n||n>e.length)&&(n=e.length);for(var t=0,r=new Array(n);t<n;t++)r[t]=e[t];return r}var S=t(264),B=h.regresar,T=h.instructionsDiv,A=h.ingredientsDiv,I=h.flexContainer,k=h.loadMoreBtn,j=h.listOfDrinks,_=h.titulo;B.style.display="none",T.style.display="none",A.style.display="none",I.style.display="none",document.querySelector(".cbox1").addEventListener("change",(function(e){e.target.checked?(document.querySelector(".flex-container").style.border="none",document.getElementById("infoBebidaDiv").style.border="none",document.querySelector(".flex-container-ddp").style.border="none",document.getElementById("user-input").style.border="none",document.querySelector("h1").classList.add("darkLayout"),document.querySelector("body").classList.add("darkLayout")):(document.querySelector(".flex-container").style.border="2px solid rgb(0, 34, 255)",document.getElementById("infoBebidaDiv").style.border="4px solid greenyellow",document.querySelector(".flex-container-ddp").style.border="4px solid rgb(255, 34, 0)",document.getElementById("user-input").style.border="4px solid rgb(0, 255, 204)",document.querySelector("h1").classList.remove("darkLayout"),document.querySelector("body").classList.remove("darkLayout"))}));var O=function(e){S("https://www.thecocktaildb.com/api/json/v1/1/search.php?s="+e).then((function(e){return e.json()})).then((function(e){if(document.getElementsByClassName("drink-type"))for(var n=document.getElementsByClassName("drink-type"),t=n.length-1;t>=0;t--)n[t].remove();var r,o=function(e,n){var t="undefined"!=typeof Symbol&&e[Symbol.iterator]||e["@@iterator"];if(!t){if(Array.isArray(e)||(t=function(e,n){if(e){if("string"==typeof e)return E(e,n);var t=Object.prototype.toString.call(e).slice(8,-1);return"Object"===t&&e.constructor&&(t=e.constructor.name),"Map"===t||"Set"===t?Array.from(e):"Arguments"===t||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(t)?E(e,n):void 0}}(e))||n&&e&&"number"==typeof e.length){t&&(e=t);var r=0,o=function(){};return{s:o,n:function(){return r>=e.length?{done:!0}:{done:!1,value:e[r++]}},e:function(e){throw e},f:o}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var i,a=!0,s=!1;return{s:function(){t=t.call(e)},n:function(){var e=t.next();return a=e.done,e},e:function(e){s=!0,i=e},f:function(){try{a||null==t.return||t.return()}finally{if(s)throw i}}}}(e.drinks);try{var i=function(){var e=r.value,n=document.createElement("p");n.innerText=e.strDrink,n.classList.add("drink-type"),j.append(n),n.addEventListener("click",(function(){!function(e,n){console.log(e.innerText),document.querySelector(".flex-container").style.display="none",document.querySelector(".busqueda").style.display="none",m.style.display="block",console.log(n,"---");var t=document.createElement("img");if(t.classList.add("class-old-image"),t.setAttribute("src",n.strDrinkThumb),document.getElementsByClassName("class-old-image"))for(var r=document.getElementsByClassName("class-old-image"),o=r.length-1;o>=0;o--)r[o].remove();if(document.getElementsByClassName("parrafo-instrucciones"))for(var i=document.getElementsByClassName("parrafo-instrucciones"),a=i.length-1;a>=0;a--)i[a].remove();var s=document.getElementsByClassName("borrar-ingredientes");if(s)for(var c=s.length-1;c>=0;c--)s[c].remove();var l=document.createElement("p");l.classList.add("parrafo-instrucciones"),l.innerText=n.strInstructions,b.append(l),t.setAttribute("id","imagen-bebida"),g.append(t),b.style.display="block",x.style.display="block",v.style.display="flex";var d=n.strIngredient1,u=n.strIngredient2,p=n.strIngredient3,f=n.strIngredient4,y=n.strIngredient5,h=n.strIngredient6,E=n.strIngredient7,S=(n.strIngredient8,n.strIngredient9,n.strIngredient10,n.strIngredient11,n.strIngredient12,n.strIngredient13,n.strIngredient14,n.strIngredient15,"".concat(null!==d?d+(null!==u?", ":"."):"").concat(null!==u?u+(null!==p?", ":"."):"").concat(null!==p?p+(null!==f?", ":"."):"").concat(null!==f?f+(null!==y?", ":"."):"").concat(null!==y?y+(null!==h?", ":"."):"").concat(null!==h?h+(null!==E?", ":"."):""));console.log(S);var B=document.createElement("p");B.classList.add("borrar-ingredientes"),B.innerText=S,x.append(B),w.innerText=n.strDrink}(n,e)})),B.addEventListener("click",(function(){document.querySelector(".flex-container").style.display="flex",document.querySelector(".busqueda").style.display="flex",B.style.display="none",document.querySelector(".class-old-image").style.display="none",document.querySelector(".parrafo-instrucciones").style.display="none",document.querySelector(".borrar-ingredientes").style.display="none",_.innerText="Pollos BAR",T.style.display="none",A.style.display="none",I.style.display="none";for(var e=document.getElementsByClassName("drink-type"),n=e.length-1;n>=0;n--)e[n].remove()}))};for(o.s();!(r=o.n()).done;)i()}catch(e){o.e(e)}finally{o.f()}console.log(e)}))},U=document.getElementById("user-input");k.addEventListener("click",(function(){O(U.value)}))})()})();
