+++ src/constants.js
@@ -1,0 +1,20 @@
+const listOfDrinks = document.getElementById('busqueda');
+const loadMoreBtn = document.getElementById('enviar');
+const regresar = document.getElementById('mostrar');
+const instructionsDiv = document.getElementById('instructions-id');
+const imagenDiv = document.getElementById('imagenDiv');
+const ingredientsDiv = document.getElementById('ingredientes-id');
+const flexContainer = document.getElementById('infoBebidaDiv');
+const titulo = document.querySelector(".bar")
+
+export const constants = {
+  listOfDrinks: listOfDrinks,
+  loadMoreBtn: loadMoreBtn,
+  regresar: regresar,
+  instructionsDiv: instructionsDiv,
+  imagenDiv: imagenDiv,
+  ingredientsDiv: ingredientsDiv,
+  flexContainer: flexContainer,
+  titulo: titulo,
+}
+
